Product CRUD and filter done

+++ Aminl Blog/src/modules/product/product.controller.js
@@ -41,6 +41,7 @@ class ProductController {
     try {
       const data = await productService.transformProductCreate(req);
       const product = await productService.createSingleRow(data);
+      console.log("product: ", product);
       res.json({
         message: "Product successfully created",
         status: "Success",
@@ -54,7 +55,10 @@ class ProductController {
 
   update = async (req, res, next) => {
     try {
+
       const data = await productService.transformProductUpdate(req);
+      // res.json(data);
+      console.log('data: ', data);
       const _id = req.params.id;
       const product = await productService.updateSingleRowByFilter(
         { _id },
@@ -76,9 +80,9 @@ class ProductController {
     try {
       const _id = req.params.id;
       const product = await productService.verifyData(_id);
-
-      const isAdmin = req.loggedInUser.UserType === "admin";
-      const isUser = product.user.toString() === req.loggedInUserId;
+      console.log("product in product Controller: ", product);
+      const isAdmin = req.loggedInUser.userType == "admin";
+      const isUser = product.user.toString() === req.loggedInUserId.toString();
       if (!(isAdmin || isUser)) {
         throw {
           message: "You cannot make changes to this product",
@@ -86,8 +90,9 @@ class ProductController {
           code: 401,
         };
       }
-
-      await cloudinaryService.deleteFile(product.image.public_id);
+      if (product.image.public_id) {
+        await cloudinaryService.deleteFile(product.image.public_id);
+      }
       const deletedProduct = await productService.deleteSingleRowByFilter({
         _id,
       });
@@ -105,11 +110,29 @@ class ProductController {
   show = async (req, res, next) => {
     try {
       const _id = req.params.id;
-      const product = await productService.verifyData(_id);
+      const product = await productService.model
+        .findOne({ _id })
+        .populate(["category", "user"]);
+      if (!product) {
+        throw {
+          message: "Product not found",
+          status: "NOT FOUND",
+          code: 404,
+        };
+      }
+      // const product = await productService.verifyData(_id);
       res.json({
         message: "Product successfully fetched",
         stauts: "Success",
-        data: product,
+        data: {
+          ...product.toObject(),
+          image: product.image.secure_url,
+          user: {
+            name: product.user.name,
+            email: product.user.email,
+            image: product.user.image.secure_url,
+          },
+        },
         options: null,
       });
     } catch (exception) {
